Cleaning up files and making test go green expect for not finding header-user-mail, which is expected
