+++ src/components/SingleArticle.jsx
@@ -28,9 +28,7 @@ const SingleArticle = () => {
       <Button data-cy="back-button" as={NavLink} to="/">
         Back
       </Button>
-      <Button data-cy='register-btn' onClick={(e) => onHandleSubmit()}>Register here!</Button>
-      {selectRegisterForm}
-      {errorMessage && (
+          {errorMessage && (
         <Container data-cy="error-message-article">
           <h1>{errorMessage}</h1>
         </Container>
